Implemented Day4,5,6,7,8,9,10

+++ AOC.Main/Results.cs
@@ -1,4 +1,4 @@
 ﻿namespace AOC
 {
-    public record Results(int Part1, int Part2);
+    public record Results(long Part1, long Part2);
 }
